order models

+++ migrations/sqls/20220714142614-products-table-up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE products (
   id serial PRIMARY KEY,
   name varchar(100) NOT NULL,
-  price int NOT NULL,
+  price numeric NOT NULL,
   category varchar(100)
 );
